pyproject.toml
[project]
name = "langgraph-rag-system"
version = "0.1.0"
description = "Advanced RAG system with Adaptive, Self, and Corrective RAG capabilities using LangGraph"
authors = [
    {name = "Tina Rostami", email = "tinarostami110@gmail.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.13,<4.0"

dependencies = [
    "beautifulsoup4>=4.14.3,<5.0.0",
    "langchain>=1.2.10,<2.0.0",
    "langgraph>=1.0.8,<2.0.0",
    "langchainhub>=0.1.21,<0.2.0",
    "langchain-community>=0.4.1,<0.5.0",
    "tavily-python>=0.7.21,<0.8.0",
    "langchain-chroma>=1.1.0,<2.0.0",
    "python-dotenv>=1.2.1,<2.0.0",
    "langchain-openai>=1.1.9,<2.0.0",
    "langchain-text-splitters>=1.1.0,<2.0.0",
    "langchain-huggingface>=1.2.0,<2.0.0",
    "sentence-transformers>=5.2.2,<6.0.0",
    "langchain-tavily>=0.2.17,<0.3.0",
    "langgraph-cli[inmem]>=0.4.12,<0.5.0",
]

[project.optional-dependencies]
dev = [
    "black>=26.1.0,<27.0.0",
    "isort>=7.0.0,<8.0.0",
    "pytest>=9.0.2,<10.0.0",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py313']

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["graph"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
